refactor: moved geo/shader nodes ui stuff into common module
