+++ geometry/__init__.py
@@ -1,15 +1,13 @@
 if "bpy" in locals():
     import importlib
-    importlib.reload(menu)
     importlib.reload(node_settings)
     importlib.reload(node_tree)
     importlib.reload(operator)
-    importlib.reload(panel)
+    importlib.reload(ui)
 else:
-    from . import menu
     from . import node_settings
     from . import node_tree
     from . import operator
-    from . import panel
+    from . import ui
 
 import bpy
